Rework integration test package hierarchy
