+++ test/integration/controllermanager/shoot/hibernation/hibernation_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shoot_test
+package hibernation_test
 
 import (
 	"context"
