+++ test/integration/controllermanager/shoot/retry/retry_suite_test.go
@@ -12,44 +12,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shoot_maintenance_test
+package retry_test
 
 import (
 	"context"
 	"testing"
 
+	"github.com/gardener/gardener/pkg/client/kubernetes"
+	"github.com/gardener/gardener/pkg/envtest"
+	"github.com/gardener/gardener/test/framework"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/sirupsen/logrus"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	logzap "sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
-
-	"github.com/gardener/gardener/pkg/client/kubernetes"
-	controllermanagerfeatures "github.com/gardener/gardener/pkg/controllermanager/features"
-	"github.com/gardener/gardener/pkg/envtest"
-	log "github.com/gardener/gardener/pkg/logger"
-	. "github.com/gardener/gardener/pkg/utils/test/matchers"
-	"github.com/gardener/gardener/test/framework"
 )
 
-func TestShootMaintenance(t *testing.T) {
-	controllermanagerfeatures.RegisterFeatureGates()
+func TestShootRetry(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Shoot Maintenance Controller Integration Test Suite")
+	RunSpecs(t, "Shoot Retry Controller Integration Test Suite")
 }
 
 var (
-	logger *logrus.Logger
-
 	ctx        = context.Background()
 	testEnv    *envtest.GardenerTestEnvironment
 	restConfig *rest.Config
 	mgrCancel  context.CancelFunc
+
 	testClient client.Client
 )
 
@@ -59,10 +51,7 @@ var _ = BeforeSuite(func() {
 	By("starting test environment")
 	testEnv = &envtest.GardenerTestEnvironment{
 		GardenerAPIServer: &envtest.GardenerAPIServer{
-			Args: []string{
-				"--disable-admission-plugins=ResourceReferenceManager,ExtensionValidator,ShootQuotaValidator,ShootValidator,ShootTolerationRestriction",
-				"--feature-gates=WorkerPoolKubernetesVersion=true",
-			},
+			Args: []string{"--disable-admission-plugins=ResourceReferenceManager,ExtensionValidator,ShootQuotaValidator,ShootValidator,ShootTolerationRestriction"},
 		},
 	}
 	var err error
@@ -72,8 +61,6 @@ var _ = BeforeSuite(func() {
 	testClient, err = client.New(restConfig, client.Options{Scheme: kubernetes.GardenScheme})
 	Expect(err).ToNot(HaveOccurred())
 
-	logger = log.AddWriter(log.NewLogger("", ""), GinkgoWriter)
-
 	By("setup manager")
 	mgr, err := manager.New(restConfig, manager.Options{
 		Scheme:             kubernetes.GardenScheme,
@@ -81,7 +68,8 @@ var _ = BeforeSuite(func() {
 	})
 	Expect(err).ToNot(HaveOccurred())
 
-	Expect(addShootMaintenanceControllerToManager(mgr)).To(Succeed())
+	err = addShootRetryControllerToManager(mgr)
+	Expect(err).ToNot(HaveOccurred())
 
 	var mgrContext context.Context
 	mgrContext, mgrCancel = context.WithCancel(ctx)
@@ -89,14 +77,9 @@ var _ = BeforeSuite(func() {
 	By("start manager")
 	go func() {
 		defer GinkgoRecover()
-		Expect(mgr.Start(mgrContext)).To(Succeed())
+		err := mgr.Start(mgrContext)
+		Expect(err).ToNot(HaveOccurred())
 	}()
-
-	By("create shoot namespace")
-	namespace := &corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{Name: "garden-dev"},
-	}
-	Expect(testClient.Create(ctx, namespace)).To(Or(Succeed(), BeAlreadyExistsError()))
 })
 
 var _ = AfterSuite(func() {
